Add new features
* Change currency to dollars
* Change amount/count to uint16_t
* Remove Game Feature GameDoubleShopSellAmount

+++ src/client/game.h
@@ -135,7 +135,7 @@ protected:
 
     // npc trade
     void processOpenNpcTrade(const std::vector<std::tuple<ItemPtr, std::string, int, int, int> >& items);
-    void processPlayerGoods(uint64_t money, const std::vector<std::tuple<ItemPtr, int> >& goods);
+    void processPlayerGoods(uint64_t money, const std::vector<std::tuple<ItemPtr, uint16_t> >& goods);
     void processCloseNpcTrade();
 
     // player trade
